Update prova.py
Added the target stations to be decided at the start (to modofy) and the timetable. timetable is a list that contein for all the agents two different arrays, once conteining all the station to travers, starting from the starting station to the target once, the other conteining the time at which the train ipotetically has to pass to the station.

Added a new 'training' method taken from the flatland examples. This is a dummy once used only to simulate the trains in the railway