Update rail_env.py
Modified the check intermediate station function. 

Idea  ----->  it's better to rewards only the trains that fully respect the order of the station to pass or to partial reward the trains that also reach part of the stations they have to reach. 
Maybe with crescent rewards based on the number of stations reached?