Added the reverse action
Added the reverse action that can be used in case of final station of the train run. A train can reverse its direction and restart in the opposit direction. 

Added the stop at stations

Added the multi-train-run possibilities. Now a train can have multiple objective to reach. Once a train run has finisched a new train run can be started